+++ src/qt/syscoinunits.h
@@ -72,7 +72,7 @@ public:
     //! Number of decimals left
     static int decimals(int unit);
     //! SYSCOIN Format as string
-    static QString format(int unit, const CAmount& amount, const uint32_t &nAsset = 0, bool plussign = false, SeparatorStyle separators = SeparatorStyle::STANDARD, bool justify = false);
+    static QString format(int unit, const CAmount& amount, const uint64_t &nAsset = 0, bool plussign = false, SeparatorStyle separators = SeparatorStyle::STANDARD, bool justify = false);
     //! Format as string (with unit)
     static QString formatWithUnit(int unit, const CAmount& amount, bool plussign=false, SeparatorStyle separators=SeparatorStyle::STANDARD);
     //! Format as HTML string (with unit)
